Added ability to expand/collapse whole subtree by holding [ctrl] to the ModelComponentBrowserTree and moved it's display elements to a separate class
