+++ modelstudio/src/com/hiveworkshop/rms/ui/gui/modeledit/ModelPanel.java
@@ -16,6 +16,7 @@ import com.hiveworkshop.rms.ui.application.viewer.AnimationController;
 import com.hiveworkshop.rms.ui.application.viewer.ControlledAnimationViewer;
 import com.hiveworkshop.rms.ui.application.viewer.perspective.PerspDisplayPanel;
 import com.hiveworkshop.rms.ui.gui.modeledit.cutpaste.ViewportTransferHandler;
+import com.hiveworkshop.rms.ui.gui.modeledit.modelcomponenttree.ModelComponentBrowserTree;
 import com.hiveworkshop.rms.ui.gui.modeledit.newstuff.listener.ModelEditorChangeNotifier;
 import com.hiveworkshop.rms.ui.gui.modeledit.selection.SelectionItemTypes;
 import com.hiveworkshop.rms.ui.gui.modeledit.selection.SelectionMode;
